requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface 
langchain-groq
sentence-transformers
python-dotenv
langgraph
langsmith
tiktoken
chromadb
pinecone
pypdf
beautifulsoup4 
requests
python-docx
unstructured
pandas